Make utils into package, create utils.plotting.
